pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "verity-gate"
version = "0.1.0"
description = "Deterministic Policy Gates for Stochastic Systems"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }

authors = [
    { name = "Ernan Hughes" }
]

dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "matplotlib",
    "seaborn",
    "sentence-transformers",
    "pyyaml",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
